Add permissions to edit the room
